requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
future==0.17.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
line-bot-sdk==1.12.1
MarkupSafe==1.1.1
Pillow==6.0.0
python-resize-image==1.1.19
requests==2.22.0
SQLAlchemy==1.3.5
urllib3==1.25.3
Werkzeug==0.15.4
